+++ CHANGELOG.md
@@ -12,8 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed 
 
-* Error when there is a comment or empty line between postings (#29) 
-* Error when there is a comment or empty line between metadata 
+* Support comment or empty line between postings (#29) 
+* Support comment or empty line between metadata 
 
 
 ## [1.0.0-alpha.11] - 2023-06-15